# Build the string directly instead of copying and then pruning tabs or newlines