--- a/src/implementation.cpp
+++ b/src/implementation.cpp
@@ -25,12 +25,9 @@ template ada::result<url_aggregator> parse<url_aggregator>(
 
 std::string href_from_file(std::string_view input) {
   // This is going to be much faster than constructing a URL.
-  std::string tmp_buffer;
   std::string_view internal_input;
   if (unicode::has_tabs_or_newline(input)) {
-    tmp_buffer = input;
-    helpers::remove_ascii_tab_or_newline(tmp_buffer);
-    internal_input = tmp_buffer;
+    internal_input = helpers::get_ascii_tab_or_newline_removed(input);
   } else {
     internal_input = input;
   }