--- a/src/url_aggregator.cpp
+++ b/src/url_aggregator.cpp
@@ -330,14 +330,9 @@ ada_really_inline void url_aggregator::parse_path(std::string_view input) {
   ada_log("url_aggregator::parse_path ", input);
   ADA_ASSERT_TRUE(validate());
   ADA_ASSERT_TRUE(!helpers::overlaps(input, buffer));
-  std::string tmp_buffer;
   std::string_view internal_input;
   if (unicode::has_tabs_or_newline(input)) {
-    tmp_buffer = input;
-    // Optimization opportunity: Instead of copying and then pruning, we could
-    // just directly build the string from user_input.
-    helpers::remove_ascii_tab_or_newline(tmp_buffer);
-    internal_input = tmp_buffer;
+    internal_input = helpers::get_ascii_tab_or_newline_removed(user_input);
   } else {
     internal_input = input;
   }