Following the comment below, I add a new function called `get_ascii_tab_or_newline_removed` to build the string directly instead of copying and then pruning tabs or newlines.

https://github.com/ada-url/ada/blob/d8f77a1d7c7928fa67ddca8ff6116bf3758f083f/src/url.cpp#L500-L502

Note:
- Due to [NRVO](https://en.cppreference.com/w/cpp/language/copy_elision) in C++ compilers, returning a local variable back to caller should be effective as pass pointers. (neither copy ctor nor move ctor will be called)
- If we add something like `inline` or `always_inline` attribute, it is better to put the definition to the header file instead of source file. Since all inline optimizitions cannot take effect if the definition cannot be reached (except link-time optimization enabled). since the definition may be in another translation unit if we put them in a source file.