@PragmaTwice 2023-10-14
> Please consider running with `ADA_BENCHMARKS=ON` (e.g., `cmake -B build -D ADA_BENCHMARKS=ON && cmake --build build`) and then execute `./build/benchmarks/benchdata` before and after your changes.

Sure. Here is the benchmark result:

Before:
```
Loading /home/twice/projects/ada/dependencies/.cache/url-dataset/out.txt
2023-10-14T10:41:21+09:00
Running ./benchmarks/benchdata
Run on (20 X 2918.4 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x10)
  L1 Instruction 32 KiB (x10)
  L2 Unified 1280 KiB (x10)
  L3 Unified 24576 KiB (x1)
Load Average: 1.05, 0.64, 0.40
ada spec: Ada follows whatwg/url
bad urls: ---------------------
ada---count of bad URLs       26
whatwg---count of bad URLs    26
-------------------------------

bytes/URL: 86.859205
curl : OMITTED
input bytes: 8688092
number of URLs: 100025
performance counters: No privileged access (sudo may help).
zuri : OMITTED
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
BasicBench_AdaURL_href              23837537 ns     23837653 ns           30 speed=364.469M/s time/byte=2.74372ns time/url=238.317ns url/s=4.19609M/s
BasicBench_AdaURL_aggregator_href   15135664 ns     15135570 ns           47 speed=574.018M/s time/byte=1.74211ns time/url=151.318ns url/s=6.6086M/s
BasicBench_whatwg                   41866413 ns     41866288 ns           16 speed=207.52M/s time/byte=4.81881ns time/url=418.558ns url/s=2.38915M/s
```

After:
```
Loading /home/twice/projects/ada/dependencies/.cache/url-dataset/out.txt
2023-10-14T10:41:26+09:00
Running ./benchmarks/benchdata-new
Run on (20 X 2918.4 MHz CPU s)
CPU Caches:
  L1 Data 48 KiB (x10)
  L1 Instruction 32 KiB (x10)
  L2 Unified 1280 KiB (x10)
  L3 Unified 24576 KiB (x1)
Load Average: 1.04, 0.65, 0.40
ada spec: Ada follows whatwg/url
bad urls: ---------------------
ada---count of bad URLs       26
whatwg---count of bad URLs    26
-------------------------------

bytes/URL: 86.859205
curl : OMITTED
input bytes: 8688092
number of URLs: 100025
performance counters: No privileged access (sudo may help).
zuri : OMITTED
--------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations UserCounters...
--------------------------------------------------------------------------------------------
BasicBench_AdaURL_href              23819007 ns     23819062 ns           29 speed=364.754M/s time/byte=2.74158ns time/url=238.131ns url/s=4.19937M/s
BasicBench_AdaURL_aggregator_href   14740361 ns     14740285 ns           47 speed=589.411M/s time/byte=1.69661ns time/url=147.366ns url/s=6.78583M/s
BasicBench_whatwg                   40722269 ns     40722100 ns           17 speed=213.351M/s time/byte=4.68712ns time/url=407.119ns url/s=2.45628M/s
```

The benchmark results are unstable (and only for x86), so I think you can run on your side if you are interested : )