@anonrig 2023-10-15
It provides a different output for me. If you run the script with `sudo`, our benchmarks enable performance counters, which provide much better output.

Before:

```
BasicBench_AdaURL_aggregator_href   17023800 ns     17022634 ns           41 GHz=3.22958 cycle/byte=6.28781 cycles/url=546.155 instructions/byte=28.6539 instructions/cycle=4.55705 instructions/ns=14.7173 instructions/url=2.48885k ns/url=169.11 speed=510.385M/s time/byte=1.95931ns time/url=170.184ns url/s=5.876M/s
```

After: 

```
BasicBench_AdaURL_aggregator_href   17118766 ns     17118780 ns           41 GHz=3.22878 cycle/byte=6.3365 cycles/url=550.383 instructions/byte=28.6567 instructions/cycle=4.52248 instructions/ns=14.6021 instructions/url=2.4891k ns/url=170.462 speed=507.518M/s time/byte=1.97037ns time/url=171.145ns url/s=5.843M/s```