@lemire 2023-10-15
I am also getting so-so results with GCC 12 and a recent Intel compiler.

Before:

```
BasicBench_AdaURL_aggregator_href   22788603 ns     22752369 ns           31 GHz=3.18841 cycle/byte=8.375 cycles/url=727.446 instructions/byte=25.1422 instructions/cycle=3.00205 instructions/ns=9.57178 instructions/url=2.18383k ns/url=228.153 speed=381.854M/s time/byte=2.6188ns time/url=227.467ns url/s=4.39625M/s
```

After:
```
BasicBench_AdaURL_aggregator_href   23516792 ns     23485895 ns           30 GHz=3.18849 cycle/byte=8.83776 cycles/url=767.641 instructions/byte=25.1306 instructions/cycle=2.84355 instructions/ns=9.06662 instructions/url=2.18282k ns/url=240.754 speed=369.928M/s time/byte=2.70323ns time/url=234.8ns url/s=4.25894M/s
```

As you can see, you save about 0.01 instructions per byte (which is zero) and the performance is down. So this PR does not seem to improve the performance.