@darowny 2024-04-30
When I was trying to improve `can_parse` I tried doing optimizations like here, especially allocation of strings and stuff, got same results, either worse performance or almost no difference